Update AQI sensor to use EPA correction formula
During the 2020 wildfire season, the EPA identified a correction formula
when using PurpleAir sensors when calculating the NowCast AQI. The
correction formula better represents the AQI during wildfire smoke
scenarios and better represents the current AQI breakpoint.

The default AQI sensor has been updated to use this corrected formula as
well as provide information on the initialization state of the sensor,
since it is supposed to be calculated using the average readings over an
hour. To provide a better experience, AQI is calculated with available
data and reports the initilization state to the sensor.

For those that don't want the corrected AQI, a new sensor is provided
using the old method of AQI calculation using the instant value of PM
2.5 atmosphere, under the name of `aqi_raw`.

Closes #2